// src/types/errors.ts
export enum DexBeeErrorCode {
  CONNECTION_FAILED = 'CONNECTION_FAILED',
  SCHEMA_VALIDATION_FAILED = 'SCHEMA_VALIDATION_FAILED',
  TRANSACTION_FAILED = 'TRANSACTION_FAILED',
  VERSION_MISMATCH = 'VERSION_MISMATCH',
  STORE_NOT_FOUND = 'STORE_NOT_FOUND',
  INDEX_NOT_FOUND = 'INDEX_NOT_FOUND',
  QUOTA_EXCEEDED = 'QUOTA_EXCEEDED',
  TIMEOUT = 'TIMEOUT',
  // Migration error codes
  MIGRATION_VALIDATION_FAILED = 'MIGRATION_VALIDATION_FAILED',
  MIGRATION_EXECUTION_FAILED = 'MIGRATION_EXECUTION_FAILED',
  MIGRATION_GENERATION_FAILED = 'MIGRATION_GENERATION_FAILED',
  MIGRATION_HISTORY_FAILED = 'MIGRATION_HISTORY_FAILED',
  MIGRATION_STATUS_FAILED = 'MIGRATION_STATUS_FAILED',
  ROLLBACK_FAILED = 'ROLLBACK_FAILED',
  ROLLBACK_VALIDATION_FAILED = 'ROLLBACK_VALIDATION_FAILED',
  DATA_TRANSFORMATION_FAILED = 'DATA_TRANSFORMATION_FAILED',
  SCHEMA_INCOMPATIBILITY = 'SCHEMA_INCOMPATIBILITY',
  DATA_INTEGRITY_VIOLATION = 'DATA_INTEGRITY_VIOLATION',
}

export class DexBeeError extends Error {
  constructor(
    public code: DexBeeErrorCode,
    message: string,
    public cause?: Error,
  ) {
    super(message)
    this.name = 'DexBeeError'
  }
}